services/tool-orchestrator/app/middleware/request_id.py
"""Middleware that ensures every request has a unique request ID."""

from __future__ import annotations

import uuid

from fastapi import Request
from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint

import structlog


class RequestIdMiddleware(BaseHTTPMiddleware):
    """Inject a request_id into structlog context and response headers."""

    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint):
        request_id = request.headers.get("X-Request-ID", str(uuid.uuid4()))
        structlog.contextvars.bind_contextvars(request_id=request_id)

        response = await call_next(request)
        response.headers["X-Request-ID"] = request_id

        structlog.contextvars.unbind_contextvars("request_id")
        return response
