services/tool-orchestrator/requirements.txt
# Tool Orchestrator service dependencies
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
httpx>=0.27.0,<1.0.0
grpcio>=1.62.0,<2.0.0
grpcio-tools>=1.62.0,<2.0.0
grpcio-health-checking>=1.62.0,<2.0.0
protobuf>=4.25.0,<6.0.0
redis[hiredis]>=5.0.0,<6.0.0
structlog>=24.1.0,<25.0.0
pybreaker>=1.2.0,<2.0.0
prometheus-client>=0.20.0,<1.0.0
